Add the start of a console application that illustrates the use of the Unit Of Work
Adds a model with two classes, a ToDo class and Category class. To illustrate using the
Unit of Work with a custom repository, there is a ToDoRepository. A generic Repository
 is used for Category model class. There is an implementation of IDbContextFactory to
create and configure the DbContext. There is then a specific derived class for DbContext
and UnitOfWork, wehich a client of this UOW needs to do.

At the moment, there is some extra code in the ToDoUnitOfWork that sould not be there.
This

There is an initial